requirements.txt
# ==============================================================================
# FRAMEWORK PRINCIPAL
# ==============================================================================
Django==5.2.1
djangorestframework==3.15.2

# ==============================================================================
# AUTENTICACIÓN Y SEGURIDAD
# ==============================================================================
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.5.0

# ==============================================================================
# BASE DE DATOS
# ==============================================================================
PyMySQL==1.1.1
mysqlclient==2.2.5

# ==============================================================================
# CONFIGURACIÓN Y VARIABLES DE ENTORNO
# ==============================================================================
python-decouple==3.8

# ==============================================================================
# DOCUMENTACIÓN DE API
# ==============================================================================
drf-yasg==1.21.7

# ==============================================================================
# TESTING Y CALIDAD DE CÓDIGO
# ==============================================================================
pytest==8.3.4
pytest-django==4.9.0
pytest-cov==6.0.0
coverage==7.6.9
factory-boy==3.3.1

# ==============================================================================
# DESARROLLO
# ==============================================================================
django-extensions==3.2.3
ipython==8.30.0

# ==============================================================================
# PRODUCCIÓN
# ==============================================================================
gunicorn==23.0.0
whitenoise==6.8.2